+++ core/ingester/src/types.d.ts
@@ -1,3 +1,24 @@
-type Options = {
+export type Options = {
   outputFile: string;
+}
+
+export type BitInfo = {
+  url?: string;
+  title?: string;
+  description?: string;
+  image_url?: string;
+}
+
+export type Bit = {
+  id?: string;
+  text?: string;
+  token_count?: number;
+  embedding?: string;
+  info?: BitInfo;
+}
+
+export type Library = {
+  version: number;
+  embedding_model: string;
+  bits: Bit[];
 }
